Moved default animation to constants
