+++ packages/core/src/common/constants.ts
@@ -3,3 +3,13 @@ export const MAX_PROGRESS = 100;
 export const MIN_PROGRESS = 0;
 export const MIN_NORM_PROGRESS = 0;
 export const MAX_NORM_PROGRESS = 1;
+export const DEFAULT_ANIMATION = {
+    in: {
+        type: 'none',
+        duration: 0
+    },
+    out: {
+        type: 'none',
+        duration: 0
+    }
+} as const;
